Updated application design.
This version removes the long-running thread model in favor
of a more simplified callback design. In this design, asyncio.sleep
is used to sleep the worker threads in the main event loop. When the
sleep function returns, a reference the the worker is returned to a
callback that (will eventually) run the worker's run method in a
thread pool and and then reschedule the sleep back in the main event
loop.